Move admin controls into a more manageable layout.
